requirements.txt
# requirements.txt
flask
feedparser
python-dotenv
google-generativeai
ollama